pyproject.toml
[project]
name = "icrl_generator"
description = "Generate ICRL thought-of-the-day images using LiteLLM, PIL, and Typer."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Michael Honaker" }
]
dynamic = ["version"]

dependencies = [
    "requests",
    "beautifulsoup4",
    "litellm",
    "pillow",
    "typer[all]",
    "tqdm",
]

[project.scripts]
icrl-gen = "icrl_generator:APP"

[tool.setuptools]
packages = ["icrl_generator"]
include-package-data = true

[tool.setuptools.package-data]
icrl_generator = ["prompts/*.txt"]

[build-system]
requires = [
    "setuptools>=69",
    "setuptools-scm",
    "wheel"
]
build-backend = "setuptools.build_meta"
